class JackknifeAB restricted to a smaller class Resample and argument cv replaced by agg_functionin MapieRegressor
