+++ mapie/utils.py
@@ -3,9 +3,7 @@ from inspect import signature
 from typing import (
     Any,
     Callable,
-    Generator,
     Iterable,
-    List,
     Optional,
     Tuple,
     Union,
@@ -14,16 +12,14 @@ from typing import (
 
 import numpy as np
 from sklearn.base import ClassifierMixin, RegressorMixin
-from sklearn.utils import resample
-from sklearn.utils.validation import _check_sample_weight, _num_samples
+from sklearn.model_selection import BaseCrossValidator
+from sklearn.utils.validation import _check_sample_weight
 
 from ._typing import ArrayLike
 
 
 def phi1D(
-    x: ArrayLike,
-    B: ArrayLike,
-    fun: Callable[[ArrayLike], ArrayLike],
+    x: ArrayLike, B: ArrayLike, fun: Callable[[ArrayLike], ArrayLike],
 ) -> ArrayLike:
     """
     The function phi1D is called by phi2D. It aims at multiplying the vector of
@@ -52,9 +48,7 @@ def phi1D(
 
 
 def phi2D(
-    A: ArrayLike,
-    B: ArrayLike,
-    fun: Callable[[ArrayLike], ArrayLike],
+    A: ArrayLike, B: ArrayLike, fun: Callable[[ArrayLike], ArrayLike],
 ) -> ArrayLike:
     """
     The function phi2D is a loop along the testing set. For each sample of the
@@ -76,199 +70,7 @@ def phi2D(
         Apply phi1D(x, B, fun) to each row x of A
 
     """
-    return np.apply_along_axis(
-        phi1D,
-        axis=1,
-        arr=A,
-        B=B,
-        fun=fun,
-    )
-
-
-class JackknifeAfterBootstrap:
-    """
-    Generate a sampling method, that resamples the training set with
-    possible bootstrap. It can replace KFold as cv argument in the MAPIE
-    class
-
-    Parameters
-    ----------
-    agg_function: str,
-        Choose among:
-        - "mean"
-        - "median"
-
-    n_resamplings : int
-        Number of resamplings
-    n_samples: int
-        Number of samples in each resampling. By default None,
-        the size of the training set
-    replace: bool
-        Wheter to replace samples in resamplings or not
-    random_states: Optional
-        List to fix random states
-
-    Attributes
-    ----------
-    split: method
-        Equivalent of KFold's split method
-    aggregate_fit: method
-        Aggregation function to determine aggregated predictions
-        on the training set
-    aggregation_predict: method
-        Aggregation function to determine aggregated predictions
-        on the testing set
-
-    Examples
-    --------
-    >>> import pandas as pd
-    >>> from mapie.utils import JackknifeAfterBootstrap
-    >>> cv = JackknifeAfterBootstrap(n_resamplings=2,random_states=[0,1])
-    >>> X = pd.DataFrame(np.array([1,2,3,4,5,6,7,8,9,10]))
-    >>> for train_index, test_index in cv.split(X):
-    ...    print(f"train index is {train_index}, test index is {test_index}")
-    train index is [5 0 3 3 7 9 3 5 2 4], test index is [8 1 6]
-    train index is [5 8 9 5 0 0 1 7 6 9], test index is [2 3 4]
-    """
-
-    valid_agg_functions_ = ["mean", "median"]
-
-    def __init__(
-        self,
-        n_resamplings: int,
-        agg_function: Optional[str] = "mean",
-        n_samples: Optional[int] = None,
-        replace: bool = True,
-        random_states: Optional[List[int]] = None,
-    ) -> None:
-
-        self.check_parameters_JackknifeAfterBoostrap(
-            agg_function=agg_function,
-            valid_agg_functions=self.valid_agg_functions_,
-            random_states=random_states,
-            n_resamplings=n_resamplings,
-        )
-        self.agg_function = agg_function
-        self.n_resamplings = n_resamplings
-        self.n_samples = n_samples
-        self.replace = replace
-        self.random_states = random_states
-
-    def check_parameters_JackknifeAfterBoostrap(
-        self,
-        agg_function: Optional[str],
-        valid_agg_functions: List[str],
-        random_states: Optional[List[int]],
-        n_resamplings: int,
-    ) -> None:
-        if not (agg_function in valid_agg_functions):
-            raise ValueError(
-                "Invalid aggregation function. "
-                "Allowed values are 'mean', 'median', and in the "
-                "last case 'proportiontocut' has to be between 0 and 1"
-            )
-        if (random_states is not None) and (
-            len(random_states) != n_resamplings
-        ):
-            raise ValueError("Incoherent number of random states")
-
-    def split(
-        self, X: ArrayLike
-    ) -> Generator[Tuple[Any, ArrayLike], None, None]:
-        """
-        Generate indices to split data into training and test set.
-
-        Parameters
-        ----------
-        X : ArrayLike of shape (n_samples, n_features)
-            Training data, where n_samples is the number of samples
-            and n_features is the number of features.
-        y : ArrayLike of shape (n_samples,)
-            The target variable for supervised learning problems.
-        groups : ArrayLike of shape (n_samples,), default=None
-            Group labels for the samples used while splitting the dataset into
-            train/test set.
-
-        Yields
-        ------
-        train : ArrayLike
-            The training set indices for that split.
-        test : ArrayLike
-            The testing set indices for that split.
-        """
-        indices = np.arange(_num_samples(X))
-        n_samples = (
-            self.n_samples if self.n_samples is not None else len(indices)
-        )
-
-        for k in range(self.n_resamplings):
-            if self.random_states is None:
-                rnd_state = None
-            else:
-                rnd_state = self.random_states[k]
-            train_index = resample(
-                indices,
-                replace=self.replace,
-                n_samples=n_samples,
-                random_state=rnd_state,
-                stratify=None,
-            )
-            test_index = np.array(
-                list(set(indices) - set(train_index)), dtype=np.int64
-            )
-            yield train_index, test_index
-
-    def aggregate_fit(self, x: ArrayLike) -> ArrayLike:
-        """
-        Take the array of predictions, made by the refitted estimators,
-        on the training set, and aggregate to produce phi-{i}(x_i) for
-        each training sample x_i
-
-        Parameters:
-        -----------
-            x : ArrayLike
-            Array of predictions, made by the refitted estimators
-
-        Returns:
-        --------
-            ArrayLike:
-            Array of phi-{i}(x_i) for each training sample x_i
-        """
-        if self.agg_function == "median":
-            return np.nanmedian(x, axis=1)
-        else:
-            return np.nanmean(x, axis=1)
-
-    def aggregate_predict(self, x: ArrayLike, k: ArrayLike) -> ArrayLike:
-        """
-        Take the array of predictions, made by the refitted estimators,
-        on the testing set, and the 1-nan array indicating for each training
-        sample which one to integrate, and aggregate to produce phi-{t}(x_t)
-        for each training sample x_t
-
-
-        Parameters:
-        -----------
-            x : ArrayLike
-                Array of predictions, made by the refitted estimators,
-                for each sample of the testing set
-            k : ArrayLike
-                1-nan array, that indicate whether to integrate the prediction
-                of a given estimator into the aggregation, for each training
-                sample
-
-        Returns:
-        --------
-                ArrayLike
-                Array of shape (testing set size,) of aggregated predictions
-                for each testing  sample
-
-        """
-        if self.agg_function == "median":
-            return phi2D(A=x, B=k, fun=lambda x: np.nanmedian(x, axis=1))
-        else:
-            K = np.where(np.isnan(k), 0.0, k)
-            return np.matmul(x, (K / (K.sum(axis=1, keepdims=True))).T)
+    return np.apply_along_axis(phi1D, axis=1, arr=A, B=B, fun=fun,)
 
 
 def check_null_weight(
@@ -427,7 +229,7 @@ def check_alpha(
 
 def check_n_features_in(
     X: ArrayLike,
-    cv: Optional[Union[float, str, JackknifeAfterBootstrap]] = None,
+    cv: Optional[Union[float, str, BaseCrossValidator]] = None,
     estimator: Optional[Union[RegressorMixin, ClassifierMixin]] = None,
 ) -> int:
     """
